feat: Add PCP system metrics and result recording
Add PCP to collect system metrics since PBench is no longer a supported
option.  Additionally, the various results produced by passmark are
stored in openmetrics.workload.passmark_$resultName

BUG: Only the first iteration is recorded into the PCP archive since the
resulting CSV file is only on the first iteration.  We should address
this in a later commit so PCP and the CSV have the subsequent iterations
available.

Co-developed-by: Cursor